refactor: reinstated Docker Desktop, goodbye Podman
The offering of Podman Desktop feels hacky; the VM also fails to work after a
simple sleep. Sometimes it never works again, and only recreating the VM fixes
it. In general a far less polished offering when compared to Docker Desktop. A
shame, as I wanted to work with their Kind integration...