refactor: plain work git credentials
Thanks to CISO and enablement teams, I cannot use a password vault to
avoid secrets in the environment. Next is the problem of rotation,
because they think its secure to force me to change the values every 15
days or less. Great job guys, I hope the audit rubber stamp was well
worth it!